Role chain attrs are set-able and don't bleed
Prior to this commit, if you had a class that "with"s a role that
"with"s another role that itself has an attribute, the object
instantiated from the class would have 2 problems with it. First, you
wouldn't be able to set the attrs by passing in values into new(). And
second, there was value-bleed between objects (similar to if "class_has"
was used).